TicketManagement/requirements.txt
fastapi
uvicorn
pymongo
passlib
python-jose[cryptography]
python-multipart
PyJWT